Tidy-up some code relating to Quadrature mouse
--HG--
branch : trunk

+++ src/iomd.h
@@ -82,7 +82,7 @@ struct iomd
         unsigned char ctrl;
         unsigned char vidcr;
         unsigned char sndstat;
-        int mousex,mousey;
+        uint16_t mousex, mousey;
 };
 
 extern struct iomd iomd;
